add more routes back to home page

+++ Bakery/Controllers/VendorsController.cs
@@ -39,7 +39,7 @@ namespace Bakery.Controllers
     }
 
     [HttpPost("/vendors/{vendorId}/orders")]
-    public ActionResult Create(int vendorId, string orderTitle, string orderDescription, string orderDate, decimal orderPrice)
+    public ActionResult Create(int vendorId, string orderTitle, string orderDescription, string orderDate, int orderPrice)
     {
       Dictionary<string, object> model = new Dictionary<string, object>();
       Vendor foundVendor = Vendor.Find(vendorId);
