+++ Bakery/Models/Order.cs
@@ -7,12 +7,12 @@ namespace Bakery.Models
     public string Description { get; set; }
     public string Title { get; set; }
     public int Id { get; }
-    public decimal Price { get; set; }
+    public int Price { get; set; }
 
     public string Date { get; set; }
     private static List<Order> _instances = new List<Order> {};
 
-    public Order(string title, string description, string date, decimal price)
+    public Order(string title, string description, string date, int price)
     {
       Title = title;
       Description = description;
